fix: address PR #18 review findings (PLAT-dxv, PLAT-buf, PLAT-zyq, PLAT-8kd, PLAT-8ws, PLAT-u8d, PLAT-v62)
- Only publish FeaturedStudentChanged when both student ID and code are
  provided, preventing empty code from being broadcast (PLAT-dxv)
- Replace time.Sleep in integration tests with OnSubscribed callback for
  deterministic subscription readiness (PLAT-buf)
- Add io.LimitReader(4KB) to Centrifugo client error response reads to
  prevent unbounded memory allocation (PLAT-zyq)
- Add connect_mode variable to Redis module, defaulting to
  PRIVATE_SERVICE_ACCESS for GKE Autopilot compatibility (PLAT-8kd)
- Add standard variables (environment, project_name, project_id, region)
  and labels to Centrifugo Terraform module (PLAT-8ws)
- Consolidate handler test mocks into shared testutil_test.go, removing
  duplicate mock implementations across test files (PLAT-u8d)
- Extract publishAsync helper to deduplicate fire-and-forget goroutine
  pattern across session handlers (PLAT-v62)
