+++ go-backend/internal/handler/session_students_test.go
@@ -10,56 +10,11 @@ import (
 	"testing"
 	"time"
 
-	"github.com/go-chi/chi/v5"
 	"github.com/google/uuid"
 
 	"github.com/jdelfino/eval/internal/auth"
 	"github.com/jdelfino/eval/internal/store"
 )
-
-// Note: mockSessionPublisher, testLogger, noopPublisher are defined in sessions_test.go
-
-// mockSessionStudentRepo implements store.SessionStudentRepository for testing.
-type mockSessionStudentRepo struct {
-	joinSessionFn        func(ctx context.Context, params store.JoinSessionParams) (*store.SessionStudent, error)
-	updateCodeFn         func(ctx context.Context, sessionID, userID uuid.UUID, code string) (*store.SessionStudent, error)
-	listSessionStudentFn func(ctx context.Context, sessionID uuid.UUID) ([]store.SessionStudent, error)
-	getSessionStudentFn  func(ctx context.Context, sessionID, userID uuid.UUID) (*store.SessionStudent, error)
-}
-
-func (m *mockSessionStudentRepo) JoinSession(ctx context.Context, params store.JoinSessionParams) (*store.SessionStudent, error) {
-	return m.joinSessionFn(ctx, params)
-}
-
-func (m *mockSessionStudentRepo) UpdateCode(ctx context.Context, sessionID, userID uuid.UUID, code string) (*store.SessionStudent, error) {
-	return m.updateCodeFn(ctx, sessionID, userID, code)
-}
-
-func (m *mockSessionStudentRepo) ListSessionStudents(ctx context.Context, sessionID uuid.UUID) ([]store.SessionStudent, error) {
-	return m.listSessionStudentFn(ctx, sessionID)
-}
-
-func (m *mockSessionStudentRepo) GetSessionStudent(ctx context.Context, sessionID, userID uuid.UUID) (*store.SessionStudent, error) {
-	return m.getSessionStudentFn(ctx, sessionID, userID)
-}
-
-func testSessionStudent() *store.SessionStudent {
-	return &store.SessionStudent{
-		ID:                uuid.MustParse("11111111-1111-1111-1111-111111111111"),
-		SessionID:         uuid.MustParse("22222222-2222-2222-2222-222222222222"),
-		UserID:            uuid.MustParse("33333333-3333-3333-3333-333333333333"),
-		Name:              "Alice",
-		Code:              "",
-		ExecutionSettings: json.RawMessage(`null`),
-		LastUpdate:        time.Date(2025, 1, 1, 0, 0, 0, 0, time.UTC),
-	}
-}
-
-func withChiParam(ctx context.Context, key, value string) context.Context {
-	rctx := chi.NewRouteContext()
-	rctx.URLParams.Add(key, value)
-	return context.WithValue(ctx, chi.RouteCtxKey, rctx)
-}
 
 // --- Join tests ---
