+++ infrastructure/terraform/modules/centrifugo/variables.tf
@@ -4,6 +4,30 @@
 # Values are provided by the calling environment.
 
 # -----------------------------------------------------------------------------
+# Common Variables (required by all modules)
+# -----------------------------------------------------------------------------
+
+variable "environment" {
+  description = "Environment name (staging or prod)"
+  type        = string
+}
+
+variable "project_name" {
+  description = "Project name for resource naming and tagging"
+  type        = string
+}
+
+variable "project_id" {
+  description = "GCP project ID"
+  type        = string
+}
+
+variable "region" {
+  description = "GCP region for all resources"
+  type        = string
+}
+
+# -----------------------------------------------------------------------------
 # Centrifugo Configuration
 # -----------------------------------------------------------------------------
